// src/api/product/dto/product-response.ts
export class CreateProductResponse {
  statusCode?: number;
  message?: string;
  data?: any;
}

export class GetProductListResponse {
  statusCode?: number;
  message?: string;
  totalCount?: number;
  data?: any;
}

export class GetProductDetailResponse {
  statusCode?: number;
  message?: string;
  data?: any;
}